KA-137 - Rename componnet FetchError to DisplayError
