+++ onboarding-task/src/components/DisplayError.tsx
